+++ packages/tag/src/Component.tsx
@@ -1,5 +1,7 @@
-import React, { ButtonHTMLAttributes, forwardRef } from 'react';
+import React, { ButtonHTMLAttributes, forwardRef, useRef } from 'react';
 import cn from 'classnames';
+import mergeRefs from 'react-merge-refs';
+import { useFocus } from '@alfalab/hooks';
 
 import styles from './index.module.css';
 
@@ -59,8 +61,20 @@ export const Tag = forwardRef<HTMLButtonElement, TagProps>(
         },
         ref,
     ) => {
+        const tagRef = useRef<HTMLButtonElement>(null);
+
+        const [focused] = useFocus(tagRef, 'keyboard');
+
         const tagProps = {
-            className: cn(styles.component, styles[size], { [styles.checked]: checked }, className),
+            className: cn(
+                styles.component,
+                styles[size],
+                {
+                    [styles.checked]: checked,
+                    [styles.focused]: focused,
+                },
+                className,
+            ),
             'data-test-id': dataTestId,
         };
 
@@ -71,7 +85,13 @@ export const Tag = forwardRef<HTMLButtonElement, TagProps>(
         };
 
         return (
-            <button ref={ref} type='button' onClick={handleClick} {...tagProps} {...restProps}>
+            <button
+                ref={mergeRefs([tagRef, ref])}
+                type='button'
+                onClick={handleClick}
+                {...tagProps}
+                {...restProps}
+            >
                 {leftAddons ? <span className={cn(styles.addons)}>{leftAddons}</span> : null}
                 <span>{children}</span>
                 {rightAddons ? <span className={cn(styles.addons)}>{rightAddons}</span> : null}